@@ -1,7 +1,7 @@
 const { ModifyEntryPlugin } = require('@angular-architects/module-federation/src/utils/modify-entry-plugin')
 const { share, withModuleFederationPlugin } = require('@angular-architects/module-federation/webpack')
 const config = withModuleFederationPlugin({
-  name: 'onecx-bookmark-ui',
+  name: 'onecx-bookmark',
   filename: 'remoteEntry.js',
   exposes: {
     './OneCXBookmarkModule': './src/main.ts'
@@ -13,8 +13,7 @@ const config = withModuleFederationPlugin({
     },
     '@angular/forms': {
       requiredVersion: 'auto',
-      includeSecondaries: true,
-      eager: false
+      includeSecondaries: true
     },
     '@angular/common': {
       requiredVersion: 'auto',
@@ -34,15 +33,18 @@ const config = withModuleFederationPlugin({
       requiredVersion: 'auto',
       includeSecondaries: true
     },
-    '@onecx/portal-integration-angular': {
+    '@ngx-translate/core': { requiredVersion: 'auto' },
+    '@onecx/accelerator': { requiredVersion: 'auto', includeSecondaries: true },
+    '@onecx/angular-accelerator': { requiredVersion: 'auto', includeSecondaries: true },
+    '@onecx/angular-auth': {
       requiredVersion: 'auto',
       includeSecondaries: true
     },
-    '@onecx/keycloak-auth': {
+    '@onecx/angular-integration-interface': {
       requiredVersion: 'auto',
       includeSecondaries: true
     },
-    '@onecx/angular-accelerator': {
+    '@onecx/angular-remote-components': {
       requiredVersion: 'auto',
       includeSecondaries: true
     },
@@ -50,13 +52,24 @@ const config = withModuleFederationPlugin({
       requiredVersion: 'auto',
       includeSecondaries: true
     },
-    '@ngx-translate/core': {
-      strictVersion: false,
-      requiredVersion: '^14.0.0'
+    '@onecx/integration-interface': {
+      requiredVersion: 'auto',
+      includeSecondaries: true
+    },
+    '@onecx/keycloak-auth': {
+      requiredVersion: 'auto',
+      includeSecondaries: true
+    },
+    '@onecx/portal-integration-angular': {
+      requiredVersion: 'auto',
+      includeSecondaries: true
+    },
+    '@onecx/portal-layout-styles': {
+      requiredVersion: 'auto',
+      includeSecondaries: true
     }
   }),
-
-  sharedMappings: ['@onecx/portal-integration-angular']
+  sharedMappings: []
 })
 
 const plugins = config.plugins.filter((plugin) => !(plugin instanceof ModifyEntryPlugin))
